@@ -1,6 +1,6 @@
 package com.buldreinfo.jersey.jaxb;
 
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.io.IOException;
 import java.util.Collection;
